+++ app/src/test/java/com/account/AppTest.java
@@ -1,7 +1,7 @@
 /*
  * This Java source file was generated by the Gradle 'init' task.
  */
-package com.account.test;
+package com.account;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
