+++ app/src/test/java/com/account/controller/impl/AccountControllerTest.java
@@ -1,4 +1,4 @@
-package com.account.test.account_tests.controller;
+package com.account.controller.impl;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
