+++ app/src/test/java/com/account/controller/impl/AccountRoleControllerTest.java
@@ -1,4 +1,4 @@
-package com.account.test.account_role_tests.controler;
+package com.account.controller.impl;
 
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
